update create_drive_form changed directory structure
